refactor(group-quiz): extract completed-result view (verbatim) [1/2]
Move the contiguous, self-contained `session.status === "completed"`
block verbatim into _components/group-quiz-result-view.tsx (owns its
own router + GroupQuiz.session i18n; takes session/tone/isHost/starting
+ restart/backToLobby/cancel callbacks). Shared PRIMARY_CTA/HERO_CTA
moved to _lib/styles.ts to avoid a page<->component cycle. Render output
identical by construction; tsc + build + 93 tests green.
page.tsx 975->861 LOC.
